fix: split settings into two to fix broken directory stuff
