+++ src/renderer/components/ErrorModal/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { setupInstallPath } from 'renderer/actions/install-path.utils';
-import settings from 'common/settings';
+import settings from 'renderer/rendererSettings';
 import { Directories } from 'renderer/utils/Directories';
 import * as fs from 'fs';
 import { Button, ButtonType } from 'renderer/components/Button';
