+++ src/renderer/components/SettingsSection/Customization.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import settings, { useSetting } from 'common/settings';
+import settings, { useSetting } from 'renderer/rendererSettings';
 import { Toggle } from '../Toggle';
 
 const SettingsItem: FC<{ name: string }> = ({ name, children }) => (
