+++ src/renderer/utils/AddonData.ts
@@ -3,7 +3,7 @@ import { GitVersions } from '@flybywiresim/api-client';
 import { Directories } from './Directories';
 import fs from 'fs';
 import { getCurrentInstall, FragmenterUpdateChecker } from '@flybywiresim/fragmenter';
-import settings from 'common/settings';
+import settings from 'renderer/rendererSettings';
 import { store } from 'renderer/redux/store';
 import { setInstalledTrack } from 'renderer/redux/features/installedTrack';
 import { setSelectedTrack } from 'renderer/redux/features/selectedTrack';
